+++ packages/saas-ui-auth-provider/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @saas-ui/auth-provider
 
+## 1.2.2
+
+### Patch Changes
+
+- eb53116: Added optional token prop to update password handler
+
+## 1.2.1
+
+### Patch Changes
+
+- ac0b682: Improve param types
+
+## 1.2.0
+
+### Minor Changes
+
+- 71b0300: Support custom auth props
+
 ## 1.1.0
 
 ### Minor Changes